Move the rest of the functionality over
