+++ src/lib/API.js
@@ -1,13 +1,252 @@
 import _ from 'underscore';
 import Ion from './Ion';
 import IONKEYS from '../IONKEYS';
-import {queueRequest, setNewOfflineStatus, onReconnect as onNetworkReconnect} from './Network';
+import {xhr, setNewOfflineStatus} from './Network';
 import CONFIG from '../CONFIG';
 import * as Pusher from './Pusher/pusher';
 import ROUTES from '../ROUTES';
+import Str from './Str';
+import Guid from './Guid';
+import redirectToSignIn from './actions/SignInRedirect';
+
+// Holds all of the callbacks that need to be triggered when the network reconnects
+const reconnectionCallbacks = [];
+
+// Queue for network requests so we don't lose actions done by the user while offline
+let networkRequestQueue = [];
+
+// Indicates if we're in the process of re-authenticating. When an API call returns jsonCode 407 indicating that the
+// authToken expired, we set this to true, pause all API calls, re-authenticate, and then use the authToken fromm the
+// response in the subsequent API calls
+let reauthenticating = false;
 
 let authToken;
 Ion.connect({key: IONKEYS.SESSION, path: 'authToken', callback: val => authToken = val});
+
+let isOffline;
+Ion.connect({
+    key: IONKEYS.NETWORK,
+    path: 'isOffline',
+    callback: (isCurrentlyOffline) => {
+        if (isOffline && !isCurrentlyOffline) {
+            _.each(reconnectionCallbacks, callback => callback());
+        }
+        isOffline = isCurrentlyOffline;
+    }
+});
+
+let credentials;
+Ion.connect({key: IONKEYS.CREDENTIALS, callback: c => credentials = c});
+
+/**
+ * @param {string} login
+ * @param {string} password
+ * @returns {Promise}
+ */
+function createLogin(login, password) {
+    // We call createLogin after getting a successful response to the Authenticate request
+    // so it's very unlikely that this will fail with 407 authToken expired which means we
+    // won't need to replay this request and thus we can use xhr instead of request
+    return xhr('CreateLogin', {
+        authToken,
+        partnerName: CONFIG.EXPENSIFY.PARTNER_NAME,
+        partnerPassword: CONFIG.EXPENSIFY.PARTNER_PASSWORD,
+        partnerUserID: login,
+        partnerUserSecret: password,
+    })
+        .then(() => Ion.set(IONKEYS.CREDENTIALS, {login, password}))
+        .catch(err => Ion.merge(IONKEYS.SESSION, {error: err}));
+}
+
+/**
+ * Adds a request to networkRequestQueue
+ *
+ * @param {string} command
+ * @param {mixed} data
+ * @returns {Promise}
+ */
+function queueRequest(command, data) {
+    return new Promise((resolve) => {
+        // Add the write request to a queue of actions to perform
+        networkRequestQueue.push({
+            command,
+            data,
+            callback: resolve,
+        });
+
+        // Try to fire off the request as soon as it's queued so we don't add a delay to every queued command
+        // eslint-disable-next-line no-use-before-define
+        processNetworkRequestQueue();
+    });
+}
+
+/**
+ * Sets API data in the store when we make a successful "Authenticate"/"CreateLogin" request
+ *
+ * @param {object} data
+ * @param {string} exitTo
+ * @returns {Promise}
+ */
+function setSuccessfulSignInData(data, exitTo) {
+    let redirectTo;
+
+    if (exitTo && exitTo[0] === '/') {
+        redirectTo = exitTo;
+    } else if (exitTo) {
+        redirectTo = `/${exitTo}`;
+    } else {
+        redirectTo = ROUTES.HOME;
+    }
+    return Ion.multiSet({
+        // The response from Authenticate includes requestID, jsonCode, etc
+        // but we only care about setting these three values in Ion
+        [IONKEYS.SESSION]: _.pick(data, 'authToken', 'accountID', 'email'),
+        [IONKEYS.APP_REDIRECT_TO]: redirectTo,
+    });
+}
+
+/**
+ * Makes an API request.
+ *
+ * For most API commands if we get a 407 jsonCode in the response, which means the authToken
+ * expired, this function automatically makes an API call to Authenticate and get a fresh authToken, and retries the
+ * original API command
+ *
+ * @param {string} command
+ * @param {mixed} data
+ * @param {string} [type]
+ * @returns {Promise}
+ */
+function request(command, data, type = 'post') {
+    // If we're in the process of re-authenticating, queue this request for after we're done re-authenticating
+    if (reauthenticating) {
+        return queueRequest(command, data);
+    }
+
+    // We treat Authenticate in a special way because unlike other commands, this one can't fail
+    // with 407 authToken expired. When other api commands fail with this error we call Authenticate
+    // to get a new authToken and then fire the original api command again
+    if (command === 'Authenticate') {
+        // yes
+        return xhr(command, data, type)
+            .then((response) => {
+                // If we didn't get a 200 response from authenticate we either failed to authenticate with
+                // an expensify login or the login credentials we created after the initial authentication.
+                // In both cases, we need the user to sign in again with their expensify credentials
+                if (response.jsonCode !== 200) {
+                    return Ion.multiSet({
+                        [IONKEYS.CREDENTIALS]: {},
+                        [IONKEYS.SESSION]: {error: response.message},
+                    })
+                        .then(redirectToSignIn);
+                }
+                setSuccessfulSignInData(response, data.exitTo);
+                return response;
+            })
+            .then((response) => {
+                // If Expensify login, it's the users first time signing in and we need to
+                // create a login for the user
+                if (data.useExpensifyLogin) {
+                    console.debug('[SIGNIN] Creating a login');
+                    createLogin(Str.generateDeviceLoginID(), Guid());
+                }
+                return response;
+            })
+            .catch(() => {
+                // If the request failed, we need to put the request object back into the queue as long as there is no
+                // doNotRetry option set in the data
+                if (data.doNotRetry !== true) {
+                    queueRequest(command, data);
+                }
+            });
+    }
+
+    // Make the http request, and if we get 407 jsonCode in the response,
+    // re-authenticate to get a fresh authToken and make the original http request again
+    // yes
+    return xhr(command, data, type)
+        .then((responseData) => {
+            if (!reauthenticating && responseData.jsonCode === 407 && data.doNotRetry !== true) {
+                reauthenticating = true;
+                return xhr('Authenticate', {
+                    useExpensifyLogin: false,
+                    partnerName: CONFIG.EXPENSIFY.PARTNER_NAME,
+                    partnerPassword: CONFIG.EXPENSIFY.PARTNER_PASSWORD,
+                    partnerUserID: credentials.login,
+                    partnerUserSecret: credentials.password,
+                    twoFactorAuthCode: ''
+                })
+                    .then((response) => {
+                        reauthenticating = false;
+
+                        // If authentication fails throw so that we hit
+                        // the catch below and redirect to sign in
+                        if (response.jsonCode !== 200) {
+                            throw new Error(response.message);
+                        }
+
+                        return response;
+                    })
+                    .then(() => xhr(command, data, type))
+                    .catch((error) => {
+                        reauthenticating = false;
+                        return Ion.multiSet({
+                            [IONKEYS.CREDENTIALS]: {},
+                            [IONKEYS.SESSION]: {error: error.message},
+                        })
+                            .then(redirectToSignIn)
+                            .then(() => Promise.reject());
+                    });
+            }
+
+            // We can end up here if we have queued up many
+            // requests and have an expired authToken. In these cases,
+            // we just need to requeue the request
+            if (reauthenticating) {
+                return queueRequest(command, data);
+            }
+            return responseData;
+        })
+        .catch(() => {
+            // If the request failed, we need to put the request object back into the queue as long as there is no
+            // doNotRetry option set in the data
+            if (data.doNotRetry !== true) {
+                queueRequest(command, data);
+            }
+        });
+}
+
+/**
+ * Process the networkRequestQueue by looping through the queue and attempting to make the requests
+ */
+function processNetworkRequestQueue() {
+    if (isOffline) {
+        // Two things will bring the app online again...
+        // 1. Pusher reconnecting (see registerSocketEventCallback at the top of this file)
+        // 2. Getting a 200 response back from the API (happens right below)
+
+        // Make a simple request every second to see if the API is online again
+        xhr('Get', {doNotRetry: true})
+            .then(() => setNewOfflineStatus(false));
+        return;
+    }
+
+    // Don't make any requests until we're done re-authenticating since we'll use the new authToken
+    // from that response for the subsequent network requests
+    if (reauthenticating || networkRequestQueue.length === 0) {
+        return;
+    }
+
+    _.each(networkRequestQueue, (queuedRequest) => {
+        request(queuedRequest.command, queuedRequest.data)
+            .then(queuedRequest.callback);
+    });
+
+    networkRequestQueue = [];
+}
+
+// Process our write queue very often
+setInterval(processNetworkRequestQueue, 1000);
 
 /**
  * When authTokens expire they will automatically be refreshed.
@@ -67,32 +306,8 @@ Pusher.registerSocketEventCallback((eventName, data) => {
     setNewOfflineStatus(isCurrentlyOffline);
 });
 
-/**
- * Sets API data in the store when we make a successful "Authenticate"/"CreateLogin" request
- *
- * @param {object} data
- * @param {string} exitTo
- * @returns {Promise}
- */
-function setSuccessfulSignInData(data, exitTo) {
-    let redirectTo;
-
-    if (exitTo && exitTo[0] === '/') {
-        redirectTo = exitTo;
-    } else if (exitTo) {
-        redirectTo = `/${exitTo}`;
-    } else {
-        redirectTo = ROUTES.HOME;
-    }
-    return Ion.multiSet({
-        // The response from Authenticate includes requestID, jsonCode, etc
-        // but we only care about setting these three values in Ion
-        [IONKEYS.SESSION]: _.pick(data, 'authToken', 'accountID', 'email'),
-        [IONKEYS.APP_REDIRECT_TO]: redirectTo,
-    });
-}
-
 // PUBLIC
+
 
 /**
  * Register a callback function to be called when the network reconnects
@@ -101,7 +316,7 @@ function setSuccessfulSignInData(data, exitTo) {
  * @param {function} cb
  */
 function onReconnect(cb) {
-    onNetworkReconnect(cb);
+    reconnectionCallbacks.push(cb);
 }
 
 /**
@@ -162,4 +377,5 @@ export {
     deleteLogin,
     getAuthToken,
     onReconnect,
+    queueRequest,
 };
