+++ src/lib/actions/Report.js
@@ -2,8 +2,7 @@ import moment from 'moment';
 import _ from 'underscore';
 import lodashGet from 'lodash.get';
 import Ion from '../Ion';
-import {queueRequest} from '../Network';
-import {onReconnect, getAuthToken} from '../API';
+import {queueRequest, onReconnect, getAuthToken} from '../API';
 import IONKEYS from '../../IONKEYS';
 import CONFIG from '../../CONFIG';
 import * as Pusher from '../Pusher/pusher';
